Proposal: use awesome pages to organize navigation
Awesome pages is a more versatile way to perfectly organize navigation.

The main difference is that instead of having all the configuration on
the mkdocs.yml file, we can have several .pages files when needed. Also
we can use regular expressions and special values like `...` that will
add the rest of the elements on the section, making it much easy for
newcomers to add new pages that do not require a special ordering.

https://pypi.org/project/mkdocs-awesome-pages-plugin/

Signed-off-by: Antonio Navarro Perez <antnavper@gmail.com>